+++ frontend/src/App.tsx
@@ -6,6 +6,7 @@ import ProtectedRoute from './auth/ProtectedRoute'
 import AdminLayout from './admin/AdminLayout'
 import CustomerManager from './admin/CustomerManager/CustomerManager'
 import ShipmentManager from './admin/ShipmentManager/ShipmentManager'
+import PackageManager from './admin/PackageManager/PackageManager'
 import './App.scss'
 
 function ChatLayout() {
@@ -34,6 +35,7 @@ function App() {
         <Route index element={<Navigate to="customers" replace />} />
         <Route path="customers" element={<CustomerManager />} />
         <Route path="shipments" element={<ShipmentManager />} />
+        <Route path="packages" element={<PackageManager />} />
       </Route>
     </Routes>
   )